+++ admin/view/admin_home.html
@@ -172,7 +172,7 @@
             <!------END OF TOTAL------>
 
             <div id="tableProd" class="recent-orders" style="display: block;">
-                <h2>Recent-orders</h2>
+
                 <table>
                     <thead>
                         <tr>
@@ -211,14 +211,14 @@
                         <label for="idProd">
                             <h3>Number Id</h3>
                         </label>
-                        <input type="text" name="idProd" id="idProd">
+                        <input type="text" name="id" id="id">
                         <span id="invalidId"></span>
                     </div>
                     <div class="itemForm">
                         <label for="nameProd">
                             <h3>Name Product</h3>
                         </label>
-                        <input type="text" name="nameProd" id="nameProd">
+                        <input type="text" name="name" id="name">
                         <span id="invalidName"></span>
                     </div>
                     <!-------->
@@ -226,7 +226,7 @@
                         <label for="imgProd">
                             <h3>Image URL</h3>
                         </label>
-                        <input type="text" name="imgProd" id="imgProd">
+                        <input type="text" name="img" id="img">
                         <span id="invalidImg"></span>
                     </div>
                     <!-------->
@@ -234,7 +234,7 @@
                         <label for="priceProd">
                             <h3>Price</h3>
                         </label>
-                        <input type="number" name="priceProd" id="priceProd">
+                        <input type="number" name="price" id="price">
                         <span id="invalidPrice"></span>
                     </div>
                     <!-------->
@@ -242,7 +242,7 @@
                         <label for="descrProd">
                             <h3>Description</h3>
                         </label>
-                        <input type="text" name="descrProd" id="descrProd">
+                        <input type="text" name="descr" id="descr">
                         <span id="invalidDesc"></span>
                     </div>
                     <!-------->
@@ -251,7 +251,7 @@
                             <h3>Type</h3>
                         </label>
                         <!-- <input type="text" name="typeProd" id="typeProd"> -->
-                        <select name="typeProd" id="typeProd">
+                        <select name="typeProd" id="type">
                             <option value="all">Tất cả</option>
                             <option value="iphone">Iphone</option>
                             <option value="samsung">Samsung</option>
